Fix: tratamento de erro ao deletar CTG
Tratamento de erro ao tentar deletar o CTG que tem registros vinculados

+++ BackEnd/src/services/ctg.service.ts
@@ -1,12 +1,13 @@
-import { PrismaClient } from "@prisma/client";
+import { PrismaClient, Prisma } from "@prisma/client";
+import AppError from '../errors/AppError';
 
 const prisma = new PrismaClient();
 
-class CTGService{
-    constructor(private prisma: PrismaClient){ }
+class CTGService {
+    constructor(private prisma: PrismaClient) { }
 
-    async criarCTG(nomeCTG: string, RTid: number){
-        try{
+    async criarCTG(nomeCTG: string, RTid: number) {
+        try {
             const ctg = await this.prisma.cTG.create({
                 data: {
                     nomeCTG: nomeCTG,
@@ -14,7 +15,7 @@ class CTGService{
                 },
             });
             return ctg.idCTG;
-        } catch (error){
+        } catch (error) {
             throw new Error("Erro ao criar CTG. Verifique os dados fornecidos.");
         }
     }
@@ -25,46 +26,64 @@ class CTGService{
             nomeCTG?: string;
             RTId?: number;
         }
-    ){
-        try{
+    ) {
+        try {
             const ctg = await this.prisma.cTG.update({
                 where: { idCTG: idCTG },
                 data: data,
             });
             return ctg;
-        }catch(error){
+        } catch (error) {
             throw new Error("Erro ao atualizar CTG. Verifique os dados fornecidos.");
         }
     }
 
-    async buscarCTGPorId(idCTG: number){
-        try{
+    async buscarCTGPorId(idCTG: number) {
+        try {
             const ctg = await this.prisma.cTG.findUnique({
                 where: { idCTG: idCTG },
             });
             return ctg;
-        }catch(error){
+        } catch (error) {
             throw new Error("Erro ao buscar CTG.");
         }
     }
 
-    async buscarCTGs(){
-        try{
+    async buscarCTGs() {
+        try {
             const ctg = await this.prisma.cTG.findMany();
             return ctg;
-        }catch(error){
+        } catch (error) {
             throw new Error("Erro ao buscar CTGs.");
         }
     }
 
-    async deletarCTG(idCTG: number){
-        try{
-            const ctg = await this.prisma.cTG.delete({
-                where: { idCTG: idCTG },
+    async deletarCTG(idCTG: number) {
+        try {
+            const ctg = await this.prisma.cTG.findUnique({
+                where: { idCTG }
             });
-            return ctg;
-        }catch(error){
-            throw new Error("Erro ao deletar CTG.");
+
+            if (!ctg) {
+                throw new AppError("CTG não encontrada.", 404);
+            }
+
+            await this.prisma.cTG.delete({
+                where: { idCTG}
+            });
+
+            return { message: "CTG deletada com sucesso." };
+
+        } catch (error: unknown) {
+            if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P2003') {
+                throw new AppError("Não é possível deletar a CTG pois ela está associado a outros registros.", 409);
+            };
+
+            if (error instanceof AppError) {
+                throw error;
+            }
+
+            throw new AppError("Erro ao deletar CTG", 500);
         }
     }
 }
